+++ packages/app-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @app/config
 
+## 0.5.0
+
+### Minor Changes
+
+- ca33836: Update to Saas UI 2.0 🥳
+
 ## 0.4.0
 
 ### Minor Changes