+++ packages/app-features/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @app/features
 
+## 13.0.0
+
+### Minor Changes
+
+- ca33836: Update to Saas UI 2.0 🥳
+
+### Patch Changes
+
+- Updated dependencies [cb6b80e]
+- Updated dependencies [807b4c8]
+- Updated dependencies [a3451a3]
+- Updated dependencies [ca33836]
+  - @saas-ui-pro/react@0.23.0
+  - @saas-ui-pro/feature-flags@0.12.0
+  - @app/config@0.5.0
+  - @saas-ui-pro/onboarding@0.12.0
+  - @app/i18n@0.4.0
+  - @ui/theme@0.6.0
+  - @ui/lib@0.5.0
+
+## 12.0.1
+
+### Patch Changes
+
+- Updated dependencies [0230795]
+  - @saas-ui-pro/react@0.22.1
+  - @ui/theme@0.5.1
+
+## 12.0.0
+
+### Minor Changes
+
+- d2e72ae: Inbox items now open the contact view for that specific notification
+- d2e72ae: Inbox list is now resizable.
+- 1e3cea0: MenuItem and Divider now imported from @chakra-ui/react
+- 66d6cf0: Update to Chakra UI 2.7 and Saas UI 2.0.0-rc.31
+- d2e72ae: Inbox pages have been moved inside the Contacts feature.
+
+### Patch Changes
+
+- Updated dependencies [e8f1af1]
+- Updated dependencies [e8f1af1]
+- Updated dependencies [66d6cf0]
+  - @saas-ui-pro/react@0.22.0
+  - @saas-ui-pro/onboarding@0.11.0
+  - @ui/theme@0.5.0
+  - @ui/lib@0.4.0
+  - @saas-ui-pro/billing@0.11.0
+
 ## 11.1.1
 
 ### Patch Changes