+++ packages/ui-lib/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ui/lib
 
+## 0.5.0
+
+### Minor Changes
+
+- ca33836: Update to Saas UI 2.0 🥳
+
+### Patch Changes
+
+- Updated dependencies [ca33836]
+  - @app/i18n@0.4.0
+
+## 0.4.0
+
+### Minor Changes
+
+- 66d6cf0: Update to Chakra UI 2.7 and Saas UI 2.0.0-rc.31
+
 ## 0.3.2
 
 ### Patch Changes