+++ packages/ui-theme/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @ui/theme
 
+## 0.6.0
+
+### Minor Changes
+
+- ca33836: Update to Saas UI 2.0 🥳
+
+### Patch Changes
+
+- Updated dependencies [ca33836]
+  - @saas-ui-pro/theme-glass@0.6.0
+
+## 0.5.1
+
+### Patch Changes
+
+- @saas-ui-pro/theme-glass@0.5.1
+
+## 0.5.0
+
+### Minor Changes
+
+- 66d6cf0: Update to Chakra UI 2.7 and Saas UI 2.0.0-rc.31
+
+### Patch Changes
+
+- Updated dependencies [66d6cf0]
+  - @saas-ui-pro/theme-glass@0.5.0
+
 ## 0.4.8
 
 ### Patch Changes