+++ saas-ui/templates/CHANGELOG.md
@@ -1,5 +1,11 @@
 # saas-ui-templates
 
+## 0.2.0
+
+### Minor Changes
+
+- ca33836: Update to Saas UI 2.0 🥳
+
 ## 0.1.0
 
 ### Minor Changes